Moved tests into 2 separate classes too
